Add a network and stack for Openbao
Openbao is an open source replacement for Hashicorp Vault. We create a
custom network for it so that Gitea can communicate with it using
OAuth2.

This change also adds an internal IP address range to Traefik. We can
prevent access to specific resources if they come from the Traefik IP
range.